--- app/src/main/java/com/young/sportsmatch/di/NetworkModule.kt
@@ -1,52 +1,0 @@
-package com.young.sportsmatch.di
-
-import com.squareup.moshi.Moshi
-import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
-import com.young.sportsmatch.BuildConfig
-import com.young.sportsmatch.data.source.remote.ApiClient
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.moshi.MoshiConverterFactory
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Singleton
-    @Provides
-    fun provideOkHttpClient(): OkHttpClient {
-        val logger = HttpLoggingInterceptor().apply {
-            level = HttpLoggingInterceptor.Level.BASIC
-        }
-
-        return OkHttpClient.Builder()
-            .addInterceptor(logger)
-            .build()
-    }
-
-    @Singleton
-    @Provides
-    fun provideRetrofit(client: OkHttpClient): Retrofit {
-        val moshi = Moshi.Builder()
-            .add(KotlinJsonAdapterFactory())
-            .build()
-
-        return Retrofit.Builder()
-            .baseUrl(BuildConfig.FIRE_BASE_URL)
-            .client(client)
-            .addConverterFactory(MoshiConverterFactory.create(moshi))
-            .build()
-    }
-
-    @Singleton
-    @Provides
-    fun provideClient(retrofit: Retrofit): ApiClient {
-        return retrofit.create(ApiClient::class.java)
-    }
-}
